+++ frontend/src/lib/components/accounts/HardwareWalletShowActionButton.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { i18n } from "$lib/stores/i18n";
-  import { showAddressAndPubKeyOnHardwareWalletProxy } from "$lib/proxy/ledger.services.proxy";
+  import { showAddressAndPubKeyOnHardwareWalletProxy } from "$lib/proxy/icp-ledger.services.proxy";
   import { busy } from "@dfinity/gix-components";
 
   const showInfo = async () =>
