+++ frontend/src/lib/proxy/icp-ledger.services.proxy.ts
@@ -2,11 +2,11 @@ import type { LedgerIdentity } from "$lib/identities/ledger.identity";
 import type {
   ConnectToHardwareWalletParams,
   RegisterHardwareWalletParams,
-} from "$lib/services/ledger.services";
+} from "$lib/services/icp-ledger.services";
 import type { Identity } from "@dfinity/agent";
 import type { NeuronInfo } from "@dfinity/nns";
 
-const importLedgerServices = () => import("../services/ledger.services");
+const importLedgerServices = () => import("../services/icp-ledger.services");
 
 export const connectToHardwareWalletProxy = async (
   callback: (params: ConnectToHardwareWalletParams) => void
