+++ frontend/src/tests/lib/api/icp-ledger.api.spec.ts
@@ -2,14 +2,14 @@ import {
   queryAccountBalance,
   sendICP,
   transactionFee,
-} from "$lib/api/ledger.api";
+} from "$lib/api/icp-ledger.api";
 import { mockMainAccount } from "$tests/mocks/accounts.store.mock";
 import { mockIdentity } from "$tests/mocks/auth.store.mock";
 import { AccountIdentifier, LedgerCanister } from "@dfinity/nns";
 import { ICPToken, TokenAmount } from "@dfinity/utils";
 import { mock } from "jest-mock-extended";
 
-describe("ledger-api", () => {
+describe("icp-ledger.api", () => {
   describe("sendICP", () => {
     let spyTransfer;
