+++ frontend/src/tests/lib/components/accounts/HardwareWalletConnectAction.spec.ts
@@ -4,7 +4,7 @@
 
 import HardwareWalletConnectAction from "$lib/components/accounts/HardwareWalletConnectAction.svelte";
 import { LedgerConnectionState } from "$lib/constants/ledger.constants";
-import { connectToHardwareWalletProxy } from "$lib/proxy/ledger.services.proxy";
+import { connectToHardwareWalletProxy } from "$lib/proxy/icp-ledger.services.proxy";
 import { mockIdentity } from "$tests/mocks/auth.store.mock";
 import en from "$tests/mocks/i18n.mock";
 import { fireEvent } from "@testing-library/dom";
